models for person and movie updated
