+++ app/src/main/java/com/wellbees/popularmovies/model/Movie.kt
@@ -1,3 +1,6 @@
 package com.wellbees.popularmovies.model
 
-data class Movie(val id: Int, val title: String, val posterPath: String, val releaseDate: String)
+class Movie(val id: Int, val title: String){
+    var posterPath : String = ""
+    var releaseDate: String = ""
+}
